# board-service/docs/swagger.yaml
basePath: /api
definitions:
...
  project-board-api_internal_dto.BoardDetailResponse:
    properties:
      assigneeId:
        type: string
      authorId:
        type: string
      boardId:
        type: string
      comments:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.CommentResponse'
        type: array
      content:
        type: string
      createdAt:
        type: string
      customFields:
        additionalProperties: true
        type: object
      dueDate:
        type: string
      participants:
        items:
          $ref: '#/definitions/project-board-api_internal_dto.ParticipantResponse'
        type: array
      projectId:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  project-board-api_internal_dto.BoardResponse:
    properties:
      assigneeId:
        type: string
      authorId:
        type: string
      boardId:
        type: string
      content:
        type: string
      createdAt:
        type: string
      customFields:
        additionalProperties: true
        type: object
      dueDate:
        type: string
      projectId:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  project-board-api_internal_dto.CommentResponse:
    properties:
      boardId:
        type: string
      commentId:
        type: string
      content:
        type: string
      createdAt:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
  project-board-api_internal_dto.CreateBoardRequest:
    properties:
      assigneeId:
        type: string
      content:
        maxLength: 5000
        type: string
      customFields:
        additionalProperties: true
        type: object
      dueDate:
        type: string
      projectId:
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - projectId
    - title
    type: object
...
  project-board-api_internal_dto.ParticipantResponse:
    properties:
      boardId:
        type: string
      createdAt:
        type: string
      id:
        type: string
      userId:
        type: string
    type: object
...
  project-board-api_internal_dto.UpdateBoardRequest:
    properties:
      assigneeId:
        type: string
      content:
        maxLength: 5000
        type: string
      customFields:
        additionalProperties: true
        type: object
      dueDate:
        type: string
      title:
        maxLength: 200
        minLength: 1
        type: string
    type: object
...
  project-board-api_internal_response.ErrorResponse:
    properties:
      error: {}
      requestId:
        type: string
    type: object
  project-board-api_internal_response.SuccessResponse:
    properties:
      data: {}
      requestId:
        type: string
    type: object
info:
  contact:
    email: support@example.com
    name: API Support
  description: |-
    프로젝트 보드 관리 시스템 API 서버입니다.
    Board, Project, Comment, Participant 관리 기능을 제공합니다.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Project Board Management API
  version: "1.0"
paths:
  /boards:
    get:
      description: 특정 Project에 속한 모든 Board를 조회합니다. 프론트엔드 호환용 엔드포인트
      parameters:
      - description: Project ID (UUID)
        in: query
        name: projectId
        required: true
        type: string
      - description: Custom Fields 필터 JSON 객체
        in: query
        name: customFields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Board 목록 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/project-board-api_internal_dto.BoardResponse'
                  type: array
              type: object
        "400":
          description: 잘못된 Project ID 또는 필터 파라미터
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project의 Board 목록 조회 (쿼리 파라미터 방식)
      tags:
      - boards
    post:
      consumes:
      - application/json
      description: 새로운 Board를 생성합니다
      parameters:
      - description: Board 생성 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project-board-api_internal_dto.CreateBoardRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Board 생성 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.BoardResponse'
              type: object
        "400":
          description: 잘못된 요청
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Board 생성
      tags:
      - boards
  /boards/{boardId}:
    delete:
      description: Board를 소프트 삭제합니다
      parameters:
      - description: Board ID (UUID)
        in: path
        name: boardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Board 삭제 성공
          schema:
            $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
        "400":
          description: 잘못된 Board ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Board를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Board 삭제
      tags:
      - boards
    get:
      description: Board ID로 상세 정보를 조회합니다 (참여자, 댓글 포함)
      parameters:
      - description: Board ID (UUID)
        in: path
        name: boardId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Board 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.BoardDetailResponse'
              type: object
        "400":
          description: 잘못된 Board ID
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Board를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Board 상세 조회
      tags:
      - boards
    put:
      consumes:
      - application/json
      description: Board 정보를 수정합니다 (제목, 내용, 단계, 중요도, 역할)
      parameters:
      - description: Board ID (UUID)
        in: path
        name: boardId
        required: true
        type: string
      - description: Board 수정 요청
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/project-board-api_internal_dto.UpdateBoardRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Board 수정 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/project-board-api_internal_dto.BoardResponse'
              type: object
        "400":
          description: 잘못된 요청
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Board를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Board 수정
      tags:
      - boards
  /boards/project/{projectId}:
    get:
      description: 특정 Project에 속한 모든 Board를 조회합니다. customFields 파라미터로 필터링 가능 (JSON
        형식)
      parameters:
      - description: Project ID (UUID)
        in: path
        name: projectId
        required: true
        type: string
      - description: 'Custom Fields 필터 JSON 객체. 예시: stage=in_progress, role=developer'
        in: query
        name: customFields
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Board 목록 조회 성공
          schema:
            allOf:
            - $ref: '#/definitions/project-board-api_internal_response.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/project-board-api_internal_dto.BoardResponse'
                  type: array
              type: object
        "400":
          description: 잘못된 Project ID 또는 필터 파라미터
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "404":
          description: Project를 찾을 수 없음
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
        "500":
          description: 서버 에러
          schema:
            $ref: '#/definitions/project-board-api_internal_response.ErrorResponse'
      summary: Project의 Board 목록 조회
      tags:
      - boards
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT 토큰을 입력하세요. 형식: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"